requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
flyerapi==1.0.8
frozenlist==1.4.1
idna==3.7
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.4
pydantic_core==2.18.4
pypika-tortoise==0.1.6
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tortoise-orm==0.20.1
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.9.4
